+++ src/Lib.hs
@@ -23,12 +23,12 @@ splitAtFirst x = fmap (drop 1) . break (x ==)
 
 -- Puts commas in numbers using the supplied decimal and comma character
 comma :: Char -> Char -> String -> Maybe String
-comma decimal com string
-    | not (checkNumber decimal string) = Nothing
-    | decimal `elem` string = Just $ (commaNoDots com $ fst split) ++ (decimal : (snd split))
+comma deci com string
+    | not (checkNumber deci string) = Nothing
+    | deci `elem` string = Just $ (commaNoDots com $ fst split) ++ (deci : (snd split))
     | otherwise = Just $ commaNoDots com string        
     where
-        split = splitAtFirst decimal string :: (String, String)
+        split = splitAtFirst deci string :: (String, String)
 
         -- Checks that a number conforms to rules:
         -- No characters other than number and decimal